delte task id and user

+++ src/tasks/tasks.service.ts
@@ -35,8 +35,8 @@ export class TasksService {
         return this.taskRepository.find();
     }
     
-    async remove(id: string): Promise<void> {
-        const result =  await this.taskRepository.delete(id);
+    async remove(id: number, user: User): Promise<void> {
+        const result =  await this.taskRepository.delete({ id, userId: user.id });
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
